一部分的VO
（吃饭去了

+++ src/vo/HotelGeneralVO.java
@@ -1,20 +1,26 @@
 package vo;
 
 public class HotelGeneralVO {
+	
 	//	酒店编号
 	public String hotelID;
+	
 	//	酒店名称
 	public String hotelName;
+	
 	//	酒店城市
 	public String city;
+	
 	//	酒店商圈
 	public String cycle;
+	
 	//	酒店星级
 	public String level;
+	
 	//	评分
 	public double score;
 	
-	public HotelGeneralVO(String hotelID, String hotelName, String city, String cycle, String level, int score) {
+	public HotelGeneralVO(String hotelID, String hotelName, String city, String cycle, String level, double score) {
 		this.hotelID = hotelID;
 		this.hotelName = hotelName;
 		this.city = city;
