+++ src/vo/MarketVO.java
@@ -4,13 +4,15 @@ public class MarketVO {
 
 	//	等级名称
 	public String marketName;
+	
 	//	等级信用度
-	public int marketCredit;
+	public double marketCredit;
+	
 	//	等级优惠
 	public double marketBenefit;
 
 	
-	public MarketVO(String marketName, int marketCredit, double d) {
+	public MarketVO(String marketName, double marketCredit, double d) {
 		this.marketName = marketName;
 		this.marketCredit = marketCredit;
 		this.marketBenefit = d;
